combined debrief & RPS
combined the debrief & RPS pages:
After they have completed the experiment, participants will see an end screen with their confirmation ID.
If they are completing for cash, they are told to email us their confirmation ID to arrange payment. After they have sent the email, they are told to press "N" to close the experiment.
If they are completing for RPS, they are told to press "Y" to complete the online RPS form.